# src/main/resources/static/api-docs.yaml
openapi: 3.0.1
info:
  title: MintNote API
  version: v0.0.1
paths:
  /notes/:
    get:
      tags:
        - Notes
      description: Get all notes
      operationId: getNotes
      responses:
        200:
          description: Notes were successfuly retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
    post:
      tags:
        - Notes
      description: Create note
      operationId: createNote
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
        required: true
      responses:
        201:
          description: Note was successfuly created
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Note'
        400:
          description: Note was not created due to invalid params
          content:
            text/plain:
              schema:
                type: string
  /notes/{id}/:
    get:
      tags:
        - Notes
      description: Get note by id
      operationId: getNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
        - Notes
      description: Delete note by id
      operationId: deleteNote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Note was successfuly deleted
          content:
            application/json:
              schema:
                type: string
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/name/:
    put:
      tags:
        - Notes
      description: Update note's name
      operationId: updateName
      requestBody:
        content:
          application/json:
            schema:
              type: string
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: Note's name was successfuly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        "404":
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/tags/:
    put:
      tags:
        - Notes
      description: Update note's name
      operationId: updateTags
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: Note's tags were successfuly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/icon/:
    put:
      tags:
        - Notes
      description: Update note's icon
      operationId: updateIcon
      requestBody:
        content:
          application/json:
            schema:
              type: string
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: Note's icon was successfuly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/parent/:
    put:
      tags:
        - Notes
      description: Update note's parent
      operationId: updateParent
      requestBody:
        content:
          application/json:
            schema:
              type: string
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: Note's parent was successfuly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/content/:
    put:
      tags:
        - Notes
      description: Update note's content
      operationId: updateContent
      requestBody:
        content:
          application/json:
            schema:
              type: string
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        200:
          description: Note's content was successfuly updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /notes/{id}/translation/{targetLanguage}/:
    get:
      tags:
        - Notes
      description: Translate note to target language
      operationId: translate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: targetLanguage
          in: path
          required: true
          schema:
            type: string
          description: Language code supported by Google Translate API
      responses:
        200:
          description: Note was successfuly translated
          content:
            text/plain:
              schema:
                type: string
        400:
          description: Error occured while translating note
          content:
            text/plain:
              schema:
                type: string
        404:
          description: Note with provided id was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
    Note:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        creationDate:
          type: string
          format: date-time
        lastModifiedDate:
          type: string
          format: date-time
        parent:
          type: string
        id:
          type: string
        content:
          type: string
    Tag:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
    ApiError:
      type: object
      properties:
        status:
          type: string
        message:
          type: string